+++ package.json
@@ -8,10 +8,13 @@
     "start": "cd frontend && yarn start",
     "lint": "cd frontend && yarn lint",
     "frontend:install": "cd frontend && yarn install",
+    "test": "npx hardhat test $(find test/unit -name '*.test.ts')",
+    "test:unit": "npx hardhat test $(find test/unit -name '*.test.ts')",
+    "test:integration": "npx hardhat test $(find test/integration -name '*.test.ts')",
     "hardhat:coverage": "hardhat coverage",
     "deploy:upgradeable": "hardhat run scripts/deploy-upgradeable.ts",
     "upgrade:contract": "hardhat run scripts/upgrade-contract.ts",
-    "test:upgradeable": "hardhat test test/UpgradeableGame.test.ts",
+    "test:upgradeable": "hardhat test test/unit/UpgradeableGame.test.ts",
     "db:migrate": "./migrations/run-migrations.sh",
     "db:clear": "./migrations/clean-db.sh",
     "fetch:game": "ts-node scripts/fetch-game.ts",
