+++ tsconfig.json
@@ -28,7 +28,8 @@
     "next-env.d.ts",
     ".next/types/**/*.ts",
     "**/*.ts",
-    "**/*.tsx"
+    "**/*.tsx",
+    ".github/workflows/seed-preview-data.ts"
   ],
   "exclude": [
     "node_modules"
